Adding support for apps running in Cluster Mode
I added a queueSettings export to the mq-config.js to begin
expanding the configurations a user can pass into the queue.
the first parameter is the "appInstanceId" which is used to
offset when an app runs. We use a combination of a random delay
for apps running in multiple servers. And then for apps running
in cluster mode we add an additional delay per cluster instance
in a 1:2 multiple.

[Ticket: X]