Refactoring, added queue house keeper, cleaned up all files
Refactored the base resources, controllers and base query to use
the latest versions. Added a queue house keeper which will check
for messages that are stuck in the queue with a batchId and not
being processed. This house keeper will remove the lock "batchId"
and allow the messages to be processed. Removed old code from refactor.

all tests passing

[Ticket: X]